--- a/composer.json
+++ b/composer.json
@@ -8,6 +8,7 @@
   "require" : {
     "xp-framework/core": "^6.10",
     "xp-framework/io-collections": "^6.5",
+    "xp-forge/mirrors": "^1.6",
     "php" : ">=5.5.0"
   },
   "bin": ["bin/xp.xp-framework.unittest.test"],