--- a/src/main/php/unittest/TestSuite.class.php
+++ b/src/main/php/unittest/TestSuite.class.php
@@ -8,7 +8,8 @@ use lang\IllegalArgumentException;
 use lang\XPClass;
 use lang\Throwable;
 use lang\Error;
-use lang\reflect\TargetInvocationException;
+use lang\mirrors\TargetInvocationException;
+use lang\mirrors\TypeMirror;
 
 /**
  * Test suite
@@ -159,17 +160,15 @@ class TestSuite extends \lang\Object {
     // "self::method" -> static method of the test class, and "method" 
     // -> the run test's instance method
     if (false === ($p= strpos($source, '::'))) {
-      return $test->getClass()->getMethod($source)->setAccessible(true)->invoke($test, $args);
+      return (new TypeMirror(get_class($test)))->method($source)->invoke($test, $args);
     }
     $ref= substr($source, 0, $p);
     if ('self' === $ref) {
-      $class= $test->getClass();
-    } else if (strstr($ref, '.')) {
-      $class= XPClass::forName($ref);
+      $mirror= new TypeMirror(get_class($test));
     } else {
-      $class= new XPClass($ref);
+      $mirror= new TypeMirror($ref);
     }
-    return $class->getMethod(substr($source, $p+ 2))->invoke(null, $args);
+    return $mirror->method(substr($source, $p + 2))->invoke(null, $args);
   }
 
   /**
@@ -181,8 +180,9 @@ class TestSuite extends \lang\Object {
    */
   protected function actionsFor($annotatable, $impl) {
     $r= [];
-    if ($annotatable->hasAnnotation('action')) {
-      $action= $annotatable->getAnnotation('action');
+    $annotations= $annotatable->annotations();
+    if ($annotations->provides('action')) {
+      $action= $annotations->named('action')->value();
       $type= XPClass::forName($impl);
       if (is_array($action)) {
         foreach ($action as $a) {
@@ -223,41 +223,41 @@ class TestSuite extends \lang\Object {
    * @throws  lang.MethodNotImplementedException
    */
   protected function runInternal($test, $result) {
-    $class= $test->getClass();
-    $method= $class->getMethod($test->name);
+    $mirror= new TypeMirror(get_class($test));
+    $method= $mirror->method($test->name);
+    $annotations= $method->annotations(); 
     $this->notifyListeners('testStarted', [$test]);
     
     // Check for @ignore
-    if ($method->hasAnnotation('ignore')) {
+    if ($annotations->provides('ignore')) {
       $this->notifyListeners('testNotRun', [
-        $result->set($test, new TestNotRun($test, new IgnoredBecause($method->getAnnotation('ignore'))))
+        $result->set($test, new TestNotRun($test, new IgnoredBecause($annotations->named('ignore')->value())))
       ]);
       return;
     }
 
     // Check for @expect
     $expected= null;
-    if ($method->hasAnnotation('expect', 'class')) {
-      $message= $method->getAnnotation('expect', 'withMessage');
-      if ('/' === $message{0}) {
-        $pattern= $message;
+    if ($annotations->provides('expect')) {
+      $expect= $annotations->named('expect')->value();
+      if (!isset($expect['withMessage'])) {
+        $expected= [XPClass::forName($expect), null]; 
+      } else if ('/' === $expect['withMessage']{0}) {
+        $expected= [XPClass::forName($expect['class']), $expect['withMessage']]; 
       } else {
-        $pattern= '/'.preg_quote($message, '/').'/';
+        $expected= [XPClass::forName($expect['class']), '/'.preg_quote($expect['withMessage'], '/').'/']; 
       }
-      $expected= [XPClass::forName($method->getAnnotation('expect', 'class')), $pattern];
-    } else if ($method->hasAnnotation('expect')) {
-      $expected= [XPClass::forName($method->getAnnotation('expect')), null];
     }
     
     // Check for @limit
     $eta= 0;
-    if ($method->hasAnnotation('limit')) {
-      $eta= $method->getAnnotation('limit', 'time');
+    if ($annotations->provides('limit')) {
+      $eta= $annotations->named('limit')->value()['time'];
     }
 
     // Check for @values
-    if ($method->hasAnnotation('values')) {
-      $annotation= $method->getAnnotation('values');
+    if ($annotations->provides('values')) {
+      $annotation= $annotations->named('values')->value();
       $variation= true;
       $values= $this->valuesFor($test, $annotation);
     } else {
@@ -267,7 +267,7 @@ class TestSuite extends \lang\Object {
 
     // Check for @actions
     $actions= array_merge(
-      $this->actionsFor($class, 'unittest.TestAction'),
+      $this->actionsFor($mirror, 'unittest.TestAction'),
       $this->actionsFor($method, 'unittest.TestAction')
     );
 
@@ -426,25 +426,27 @@ class TestSuite extends \lang\Object {
    * mark all tests in this class as skipped and continue with tests from
    * other classes (if available)
    *
-   * @param  lang.XPClass class
+   * @param  lang.mirrors.TypeMirror mirror
    * @return void
    */
-  protected function beforeClass($class) {
-    foreach ($class->getMethods() as $m) {
-      if (!$m->hasAnnotation('beforeClass')) continue;
+  protected function beforeClass($mirror) {
+    foreach ($mirror->methods() as $method) {
+      if (!$method->annotations()->provides('beforeClass')) continue;
       try {
-        $m->invoke(null, []);
+        $method->invoke(null, []);
       } catch (TargetInvocationException $e) {
         $cause= $e->getCause();
         if ($cause instanceof PrerequisitesNotMetError) {
           throw $cause;
         } else {
-          throw new PrerequisitesNotMetError('Exception in beforeClass method '.$m->getName(), $cause);
+          throw new PrerequisitesNotMetError('Exception in beforeClass method '.$method->name(), $cause);
         }
       }
     }
-    foreach ($this->actionsFor($class, 'unittest.TestClassAction') as $action) {
-      $action->beforeTestClass($class);
+
+    $type= $mirror->type();
+    foreach ($this->actionsFor($mirror, 'unittest.TestClassAction') as $action) {
+      $action->beforeTestClass($type);
     }
   }
   
@@ -452,17 +454,19 @@ class TestSuite extends \lang\Object {
    * Call afterClass methods of the last test's class. Ignore any 
    * exceptions thrown from these methods.
    *
-   * @param  lang.XPClass class
+   * @param  lang.mirrors.TypeMirror mirror
    * @return void
    */
-  protected function afterClass($class) {
-    foreach ($this->actionsFor($class, 'unittest.TestClassAction') as $action) {
-      $action->afterTestClass($class);
+  protected function afterClass($mirror) {
+    $type= $mirror->type();
+    foreach ($this->actionsFor($mirror, 'unittest.TestClassAction') as $action) {
+      $action->afterTestClass($type);
     }
-    foreach ($class->getMethods() as $m) {
-      if (!$m->hasAnnotation('afterClass')) continue;
+
+    foreach ($mirror->methods() as $method) {
+      if (!$method->annotations()->provides('afterClass')) continue;
       try {
-        $m->invoke(null, []);
+        $method->invoke(null, []);
       } catch (TargetInvocationException $ignored) { }
     }
   }
@@ -476,8 +480,8 @@ class TestSuite extends \lang\Object {
    * @throws  lang.MethodNotImplementedException in case given argument is not a valid testcase
    */
   public function runTest(TestCase $test) {
-    $class= $test->getClass();
-    if (!$class->hasMethod($test->name)) {
+    $mirror= new TypeMirror(get_class($test));
+    if (!$mirror->methods()->provides($test->name)) {
       throw new MethodNotImplementedException('Test method does not exist', $test->name);
     }
     $this->notifyListeners('testRunStarted', [$this]);
@@ -485,9 +489,9 @@ class TestSuite extends \lang\Object {
     // Run the single test
     $result= new TestResult();
     try {
-      $this->beforeClass($class);
+      $this->beforeClass($mirror);
       $this->runInternal($test, $result);
-      $this->afterClass($class);
+      $this->afterClass($mirror);
       $this->notifyListeners('testRunFinished', [$this, $result, null]);
     } catch (PrerequisitesNotMetError $e) {
       $this->notifyListeners('testSkipped', [$result->setSkipped($test, $e, 0.0)]);
@@ -509,11 +513,11 @@ class TestSuite extends \lang\Object {
     $result= new TestResult();
     try {
       foreach ($this->sources as $classname => $groups) {
-        $class= new XPClass($classname);
+        $mirror= new TypeMirror($classname);
 
         // Run all tests in this class
         try {
-          $this->beforeClass($class);
+          $this->beforeClass($mirror);
         } catch (PrerequisitesNotMetError $e) {
           foreach ($groups as $group) {
             foreach ($group->tests() as $test) {
@@ -528,7 +532,7 @@ class TestSuite extends \lang\Object {
             $this->runInternal($test, $result);
           }
         }
-        $this->afterClass($class);
+        $this->afterClass($mirror);
       }
       $this->notifyListeners('testRunFinished', [$this, $result, null]);
     } catch (StopTests $stop) {