--- a/src/test/php/unittest/tests/SuiteTest.class.php
+++ b/src/test/php/unittest/tests/SuiteTest.class.php
@@ -356,17 +356,6 @@ class SuiteTest extends TestCase {
   }
 
   #[@test]
-  public function native_exceptions_make_test_fail() {
-    $test= newinstance(TestCase::class, ['fixture'], [
-      '#[@test] fixture' => function() { throw new \Exception('Test'); }
-    ]);
-    $this->assertInstanceOf(
-      'lang.XPException',
-      $this->suite->runTest($test)->failed[$test->hashCode()]->reason
-    );
-  }
-
-  #[@test, @action(new RuntimeVersion('>=7.0.0'))]
   public function native_php7_errors_make_test_fail() {
     $test= newinstance(TestCase::class, ['fixture'], [
       '#[@test] fixture' => function() { $null= null; $null->method(); }