Supersedes #5 
## Performance

Before:

``` sh
Timm@slate ~/devel/xp/unittest [master]
$ xp test src/test/php/
# ...

♥: 388/399 run (11 skipped), 388 succeeded, 0 failed
Memory used: 5220.40 kB (5284.60 kB peak)
Time taken: 0.126 seconds
```

``` sh
Timm@slate ~/devel/xp/core [master]
$ xp test src/test/config/unittest/core.ini
# ...

♥: 2132/2161 run (29 skipped), 2132 succeeded, 0 failed
Memory used: 9823.09 kB (10567.55 kB peak)
Time taken: 1.556 seconds
```

After, and including use of https://github.com/xp-forge/mirrors/pull/34:

``` sh
Timm@slate ~/devel/xp/unittest [refactor/use-mirrors]
$ xp test src/test/php/
# ...

♥: 388/399 run (11 skipped), 388 succeeded, 0 failed
Memory used: 6155.25 kB (7908.18 kB peak)
Time taken: 0.211 seconds
```

``` sh
Timm@slate ~/devel/xp/core [master]
$ xp test src/test/config/unittest/core.ini
# ...

♥: 2132/2161 run (29 skipped), 2132 succeeded, 0 failed
Memory used: 12420.11 kB (12897.98 kB peak)
Time taken: 1.625 seconds
```

_Before the refactoring to xp-forge/mirrors, the figures are  7867.74 kB (8337.23 kB peak) and 0.518 seconds!_